bugfix(multimedia) - Fixes after CR
